requirements.txt
aioredis==1.3.1
altgraph==0.17.2
amqp==5.0.6
asgiref==3.4.1
aspose-words-cloud==22.2.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.2.1
Automat==20.2.0
billiard==3.6.4.0
boto3==1.20.13
botocore==1.23.13
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.4
celery==5.1.2
cffi==1.15.0
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.26.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
cycler==0.10.0
daphne==3.0.2
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.6
django-appconf==1.0.5
django-background-task==0.1.8
django-cors-headers==3.8.0
django-crontab==0.7.1
django-debug-toolbar==3.2.4
django-environ==0.7.0
django-oauth-toolkit==1.5.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-storages==1.12.3
django-user-accounts==3.0.4
djangorestframework==3.12.4
# djangorestframework-simplejwt==4.4.0
docx==0.2.4
docx-utils==0.1.3
drf-social-oauth2==1.0.9
fcm-django==1.0.5
ffmpy==0.3.0
fire==0.4.0
firebase-admin==5.0.3
fonttools==4.27.1
# google-api-core==2.2.0
google-api-python-client==2.28.0
google-auth==2.3.2
google-auth-httplib2==0.1.0
google-cloud-core==2.1.0
pycparser==2.21
pycrypto==2.6.1
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
hiredis==2.0.0
httplib2==0.20.1
hyperlink==21.0.0
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.5
# importlib-metadata==4.8.1
incremental==21.3.0
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
jwcrypto==1.0
jwt==1.3.1
kiwisolver==1.3.1
kombu==5.1.0
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.3.4
moviepy==1.0.3
msgpack==0.6.2
numpy==1.19.5
oauthlib==3.1.1
openapi-codec==1.3.2
# opencv-python==4.5.3.56
packaging==21.0
pdf2docx==0.5.2
Pillow==8.3.2
proglog==0.1.9
prompt-toolkit==3.0.20
proto-plus==1.19.6
protobuf==3.19.0
# psycopg2==2.7.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycrypto==2.6.1
pyinstaller==4.9
pyinstaller-hooks-contrib==2022.0
# PyJWT==2.3.0
# PyMuPDF==1.19.4
pyOpenSSL==22.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-docx==0.8.11
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
# redis-server==6.0.9
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
s3transfer==0.5.0
service-identity==21.1.0
simplejson==3.17.5
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
termcolor==1.1.0
tinify==1.5.2
tqdm==4.62.2
twilio==7.3.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.6.0
zope.interface==5.4.0
